File I/O

File I/O

WaterModels.parse_epanet — Function.

parse_epanet(path)

Parses an EPANET (.inp) file from the file path path and returns a WaterModels data structure (a dictionary of data). See the OpenWaterAnalytics Wiki for a thorough description of the EPANET format and its components.

source

WaterModels.parse_file — Function.

parse_file(path)

Parses an EPANET (.inp) or JavaScript Object Notation (JSON) file from the file path path, depending on the file extension, and returns a WaterModels data structure (a dictionary of data).

source

WaterModels.parse_json — Function.

parse_json(path)

Parses a JavaScript Object Notation (JSON) file from the file path path and returns a WaterModels data structure (a dictionary of data).

source