Variables

Variables

The following methods provide a compositional approach for defining common variables used in water network optimization models. These methods are always defined over AbstractWaterModel.

WaterModels.variable_head — Method.

Creates bounded (by default) or unbounded total hydraulic head variables for all nodes in the network, i.e., h[i] for i in node.

source

WaterModels.variable_head_bounded — Method.

Creates bounded total hydraulic head variables for all nodes in the network, i.e., h[i] for i in node.

source

WaterModels.variable_head_gain — Method.

Creates head gain variables corresponding to all pumps in the network, i.e., g[a] for a in pump. These denote head gains between nodes i and j.

source

WaterModels.variable_head_unbounded — Method.

Creates unbounded total hydraulic head variables for all nodes in the network, i.e., h[i] for i in node. Note that head variables corresponding to reservoir and tank component types are also unbounded, here.

source

WaterModels.variable_head_violation — Method.

Creates variables denoting absolute constraint violations with respect to head bounds for all nodes in the network, i.e., h_viol[i] for i in node.

source

WaterModels.variable_reservoir — Method.

Creates outgoing flow variables for all reservoirs in the network, i.e., qr[i] for i in reservoir. Note that these variables are always nonnegative, as there is never incoming flow to a reservoir.

source

WaterModels.variable_tank — Method.

Creates outgoing flow variables for all tanks in the network, i.e., qt[i] for i in tank. Note that, unlike reservoirs, tanks can have inflow.

source

WaterModels.variable_volume — Method.

Creates bounded (by default) or unbounded (but still nonnegative) volume variables for all tanks in the network, i.e., V[i] for i in tank.

source

WaterModels.variable_volume_bounded — Method.

Creates bounded volume variables for all tanks in the network, i.e., V[i] for i in tank. Bounds are related to minimum and maximum tank levels.

source

WaterModels.variable_volume_unbounded — Method.

Creates unbounded volume variables for all tanks in the network, i.e., V[i] for i in tank. Note that since volume is inherently a nonnegative quantity, however, these volume variables still contain lower bounds of zero.

source